feat: fixing bugs on usecase tenggat waktu and reminder
