+++ FrontEnd/src/Components/Button/SirioDatePickerCustomInput.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import SirioButton from "../Button/SirioButton";
+import SirioAxiosBase from '../../Services/SirioAxiosBase';
 
 class SirioDatePickerCustomInput extends Component {
 
@@ -8,7 +9,7 @@ class SirioDatePickerCustomInput extends Component {
         if (this.props.unchangedContent) {
             content = this.props.children;
         } else {
-            content = this.props.value ? this.props.value : this.props.children;
+            content = this.props.value ? SirioAxiosBase.formatDateFromSirioDatePicker(this.props.value) : this.props.children;
         }
 
         return (
