+++ FrontEnd/src/Components/Tables/Rekomendasi/TabelRekomendasi.jsx
@@ -5,6 +5,7 @@ import SirioTable from '../SirioTable';
 import RekomendasiService from '../../../Services/RekomendasiService';
 import { NavLink } from 'react-router-dom';
 import classes from './TabelRekomendasi.module.css';
+import SirioAxiosBase from '../../../Services/SirioAxiosBase';
 
 /**
  * Kelas untuk membuat komponen tabel rekomendasi
@@ -28,14 +29,8 @@ export default class TabelRekomendasi extends React.Component {
     async renderRows() {
         const response = await RekomendasiService.getRekomendasiByLoggedInUser();
 
-        var fetchedRows = [];
-        response.data.result.map((entry, i) => {
-            entry.no = i + 1;
-            return fetchedRows.push(entry);
-        })
-
         this.setState({
-            rowList: fetchedRows
+            rowList: response.data.result
         })
     }
 
@@ -60,7 +55,6 @@ export default class TabelRekomendasi extends React.Component {
             headerStyle: (colum, colIndex) => {
                 return { width: "25%", textAlign: 'left' };
             }
-
         }, {
             dataField: 'status',
             text: 'STATUS',
@@ -146,20 +140,20 @@ export default class TabelRekomendasi extends React.Component {
 
     static aturTenggatWaktu(date, id) {
         // yyyy-MM-dd
-        let month = (date.getMonth() + 1);
-        if (month < 10) {
-            month = "0" + month;
-        }
-        let simpleDate = date.getDate();
-        if (simpleDate < 10) {
-            simpleDate = "0" + simpleDate;
-        }
-        let formattedDate = date.getFullYear() + "-" + month + "-" + simpleDate;
+        // let month = (date.getMonth() + 1);
+        // if (month < 10) {
+        //     month = "0" + month;
+        // }
+        // let simpleDate = date.getDate();
+        // if (simpleDate < 10) {
+        //     simpleDate = "0" + simpleDate;
+        // }
+        // let formattedDate = date.getFullYear() + "-" + month + "-" + simpleDate;
+        const newDate = [(date.getFullYear()), (date.getMonth() + 1), date.getDate()];
         RekomendasiService.setTenggatWaktu({
             id: id,
-            tenggatWaktu: formattedDate
+            tenggatWaktuDate: newDate
         }).then(response => console.log(response));
-        //TODO: render ulang
     }
 
     // Formatter untuk render button kedua
@@ -191,7 +185,7 @@ export default class TabelRekomendasi extends React.Component {
                     id={row.id}
                     handleChange={(date, id) => TabelRekomendasi.aturTenggatWaktu(date, id)}
                 >
-                    {tenggatWaktuExist ? tenggatWaktu : "Tenggat Waktu"}
+                    {SirioAxiosBase.formatDateFromString(tenggatWaktu)}
                 </SirioDatePickerButton>
             )
         } else if (text) {
@@ -200,7 +194,7 @@ export default class TabelRekomendasi extends React.Component {
                     purple
                     text
                 >
-                    {tenggatWaktu}
+                    {SirioAxiosBase.formatDateFromString(tenggatWaktu)}
                 </SirioButton>
             )
         } else if (disabled) {
